adding images & changing the place of the export
