+++ RES/src/app/dashboard/components/list/list.component.html
@@ -3,27 +3,6 @@
   (afterCollapse)="disPatchSetInView(true)"
   (afterExpand)="disPatchSetInView(false)"
 >
-  <button
-    mat-icon-button
-    [matMenuTriggerFor]="menu"
-    aria-label="Options for exporting files"
-  >
-    <mat-icon>more_vert</mat-icon>
-  </button>
-  <mat-menu #menu="matMenu">
-    <button mat-menu-item (click)="exportFile('pdf')">
-      <mat-icon>dialpad</mat-icon>
-      <span>Pdf</span>
-    </button>
-    <button mat-menu-item (click)="exportFile('docx')">
-      <mat-icon>voicemail</mat-icon>
-      <span>word</span>
-    </button>
-    <button mat-menu-item (click)="exportFile('xlsx')">
-      <mat-icon>notifications_off</mat-icon>
-      <span>Excel</span>
-    </button>
-  </mat-menu>
   <div
     class="prevent-scroll"
     *ngIf="!isPaginatedList"
