+++ RES/src/app/dashboard/components/list/paginated-list/filter-paginated-list/filter-paginated-list.component.scss
@@ -1,0 +1,8 @@
+.text-right {
+  text-align: right;
+}
+
+img {
+  height: 30px;
+  width: 30px;
+}
